refactor for generic building quizzes views
